requirements.txt
attrs==17.4.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
dask==0.16.0
decorator==4.1.2
elasticsearch==2.3.0
filetype==1.0.0
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
idna==2.6
image-match==1.1.2
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
matplotlib==2.1.1
msgpack-python==0.5.1
networkx==2.0
numpy==1.13.3
Pillow==5.0.0
pluggy==0.6.0
py==1.5.2
pymongo==3.6.0
pyparsing==2.2.0
pytest==3.3.2
pytest-tornado==0.4.5
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
pyzmq==16.0.3
requests==2.18.4
requests-toolbelt==0.8.0
scikit-image==0.12.3
scipy==1.0.0
six==1.11.0
toolz==0.9.0
tornado==4.5.2
torndb==0.3
urllib3==1.22
Werkzeug==0.14.1
